Refactor backend into modular routes
