+++ backend/services/stripe.js
@@ -1,0 +1,5 @@
+const stripe = require('stripe')(process.env.STRIPE_SECRET_KEY, {
+  apiVersion: '2022-08-01',
+});
+
+module.exports = stripe;
